pyproject.toml
[project]
name = "budget-app"
version = "0.1.0"
description = "Personal finance tracker with AI-powered categorization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "anthropic>=0.42.0",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.18",
    "plaid-python>=27.0.0",
    "aiofiles>=24.1.0",
    "cryptography>=44.0.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.11.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
